Develop

+++ src/main/java/com/vroommates/VroomMates/model/usermodel/dto/UserResponseDTO.java
@@ -16,6 +16,6 @@ public class UserResponseDTO {
     private Double lon;
     private String pfp;
     private LocalDateTime createdAt;
-    private Integer distance;
-    private Integer co2;
+    private Double distance;
+    private Double co2;
 }
